Updated to 1.1

+++ src/com/matcracker/PMManagerServers/Settings/PMServersManager.java
@@ -3,8 +3,11 @@ package com.matcracker.PMManagerServers.settings;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
+
+import com.matcracker.PMManagerServers.API.APIManager;
 import com.matcracker.PMManagerServers.API.UtilityServersAPI;
 import com.matcracker.PMManagerServers.lang.BaseLang;
+import com.matcracker.PMManagerServers.utility.FileChooser;
 import com.matcracker.PMManagerServers.utility.Utility;
 import com.matcracker.PMManagerServers.utility.UtilityColor;
 
@@ -57,28 +60,42 @@ public class PMServersManager {
 		
 		serverManagerMenu();
 	}
-
+	
 	private static void deleteServer(int server) throws IOException {
 		int nservers = UtilityServersAPI.getNumberServers();
+		int npath = UtilityServersAPI.getNumberPaths();
 		
-		if(nservers > 1){
+		if(nservers > 1 && npath > 1){
 			Files.delete(new File("ServersName" + File.separator + "ServerName_" + server + ".pm").toPath());
+			Files.delete(new File("Path" + File.separator + "path_" + server + ".pm").toPath());
 			for(int i = server+1; i <= nservers; i++){
 				String temp = UtilityServersAPI.getNameServer(i);
+				String tempPath = UtilityServersAPI.getPath(i);
+				
 				Files.delete(new File("ServersName" + File.separator + "ServerName_" + i + ".pm").toPath());
+				Files.delete(new File("Path" + File.separator + "path_" + i + ".pm").toPath());
+				
 				UtilityServersAPI.setNameServer(i-1, temp);
+				UtilityServersAPI.setPath(i-1, tempPath);
 			}
 		
 			nservers--;
 			UtilityServersAPI.setNumberServer(nservers);
 			Utility.waitConfirm(UtilityColor.COLOR_GREEN + BaseLang.translate("pm.serverManager.correctDelete"));
 		}else
-			Utility.waitConfirm("Can't delete this server! It's the last!");
+			Utility.waitConfirm(BaseLang.translate("pm.serverManager.cantDelete"));
 	
 	}
 	
 	private static void addServer(){
 		String newName = Utility.readString(BaseLang.translate("pm.serverManager.choose") + " ", "[" + BaseLang.translate("pm.serverManager.suggest") + " " + UtilityServersAPI.getDefaultServerName() + "]");
+		String newPath = "";
+		
+		if(APIManager.isServerMode())
+			newPath = Utility.readString("Select path for this server: ", "[Example: /home/User/PocketMine-MP/]");
+		else
+			newPath = FileChooser.getPhar("Select new path of a server");
+		
 		int nservers = UtilityServersAPI.getNumberServers();
 		
 		nservers++;
@@ -86,24 +103,36 @@ public class PMServersManager {
 			newName = UtilityServersAPI.getDefaultServerName() + "_" + nservers;
 		
 		UtilityServersAPI.setNameServer(nservers, newName);
+		
+		if(newPath != null)
+			UtilityServersAPI.setPath(nservers, newPath);
+		
 		UtilityServersAPI.setNumberServer(nservers);
 		
 		Utility.waitConfirm(UtilityColor.COLOR_GREEN + BaseLang.translate("pm.serverManager.correctAdd"));
 	}
 
-	private static void changeServerName(int server) {
+	private static void changeServerName(int server){
 		String newName = Utility.readString(BaseLang.translate("pm.serverManager.choose") + " ", "[" + BaseLang.translate("pm.serverManager.suggest") + " " + UtilityServersAPI.getDefaultServerName() + "]");
+		String newPath = "";
+		
 		File oldBackupFile = new File("Backups" + File.separator + "Servers" + File.separator + UtilityServersAPI.getNameServer(server) + ".zip");
 		
 		if(newName.equalsIgnoreCase(""))
 			newName = UtilityServersAPI.getDefaultServerName() + "_" + server;
 		
+		if(APIManager.isServerMode())
+			newPath = Utility.readString("Select path for this server: ", "[Example: /home/User/PocketMine-MP/]");
+		else
+			newPath = FileChooser.getPhar("Select new path of a server");
+		
 		File newBackupFile = new File("Backups" + File.separator + "Servers" + File.separator + newName + ".zip");
 		UtilityServersAPI.setNameServer(server, newName);
-
+		UtilityServersAPI.setPath(server, newPath);
+		
 		if(oldBackupFile.exists())
 			oldBackupFile.renameTo(newBackupFile);
 		
 		Utility.waitConfirm(UtilityColor.COLOR_GREEN + BaseLang.translate("pm.serverManager.correctChange"));
-	}
+	}	
 }
